% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectGBM.R
\name{TrenaProjectGBM}
\alias{TrenaProjectGBM}
\title{Define an object of class TrenaProjectPlacneta}
\usage{
TrenaProjectGBM(quiet = TRUE)
}
\value{
An object of the TrenaProjectGBM class
}
\description{
Expression, variant and covariate data for the 21 GBM loci
}
